raspberry/requirements.txt
RPi.GPIO==0.7.1
requests==2.31.0
python-dateutil==2.8.2